split hydraulic_discharge.hpp into h and cpp files
